* removed dependency for JPath in MVC classes * "normalized" Controller interface:  - added getLayout function (and corresponding property)  - added getLayoutName function  - renamed: getViews to getViewNames  - added: getViews [returns a list of the view info/JSON objects]  - renamed: getPartials to getPartialNames  - added: getPartials [returns a list of the partial info/JSON objects]
* added ANT file parser.xml for parsing template files (*.ehtml):
layouts, views, partials (as listed in directories.json)
* renamed JavaScript files for ANT build tasks:
  * InitForAntRhinoScriptEnv.js -> InitRhinoEnv.js
  * AntScriptFileHandler.js -> AntFileHandler.js
  * InitForAntNodeJsScriptEnv.js -> InitNodeJsEnv.js
  * InitForAntDefaultScriptEnv.js -> InitAntDefaultEnv.js

* project-settings: set JavaScript-parsing to assets/www/ (i.e. exclude
other folders, e.g. build/...)

* added transitions (animations) for views: now jQuery Mobile page
transitions can be used in render(..., {transition: 'jqm transition
name', reverse: [true|false]})
  - default transition now set in PresentationManager (to: 'none'),
instead of in main.js::afterLoadingControllers()
  - added example dialogDescriptionSCXML.xml that uses transitions
* modified language-menu handling (now handling all user actions via
events/DialogEngine)
* constants: 
  - added alternative String-type as argument in setBasePath (for
explicitly setting a base path)
  - added function for getting directories.json-path

* added "hidden" handling for (optional) @argument-parameter in
performHelper implementation

* BUGFIX for @render handling
* FIX @argument is now treated as "local" variable (i.e. added
collision/scope handling)
* BUGFIX for @argument processing during parsing of templates
* BUGFIX template parsing: avoid "false negatives" (i.e. avoid error
messages for non-errors)
* BUGFIX semantic interpreter: apply stopword function to encoded
string!

* updated test page for grammar for new multi-language support
* added optional argument in GrammarConverter load-grammar function

+++ assets/www/mmirf/tools/build/ant/AntFileHandler.js
@@ -27,7 +27,14 @@
 
 /*
  * This JavaScript is expected to run in the Rhino environment 
- * (i.e. with access to Java classes).
+ * (or the default ANT JavaScript environment)
+ * 
+ * -> requires access to Java classes
+ * 
+ * 
+ * SETTINGS:
+ *  - isDebugOutput: disable debugging output (console) by setting global variable isDebugOutput to FALSE
+ * 
  */
 
 importClass(java.io.File);
@@ -35,11 +42,19 @@ importClass(java.io.FileReader);
 importClass(java.io.FileWriter);
 importClass(java.io.LineNumberReader);
 
-function loadLocalFile(path){
+
+var isDebugOutput = typeof isDebugOutput !== 'undefined'? isDebugOutput : true; 
+
+/*
+ * see loadLocalFile in IFileHandler.js
+ */
+function loadLocalFile(path, type){
 	
 	var f = new File(path);
 	
-	console.log('reading: json file exists? '+f.exists());
+	if(f.exists() === false){
+		throw new Error('error reading file '+path+': file does not exist!');
+	}
 	
 	var r = new LineNumberReader(new FileReader(f));
 	
@@ -50,12 +65,15 @@ function loadLocalFile(path){
 	}
 	r.close();
 	
-	console.log('read contents from file: '+path);
+	if(isDebugOutput) console.log('read contents from file: '+path);
 	
 	return JSON.parse(theJSONgrammarString);
 }
 
-function saveToFile(str, path){
+/*
+ * see saveToFile in IFileHandler.js
+ */
+function saveToFile(str, path, doNotOverWrite){
 	
 	if(typeof str !== 'string'){
 		str = JSON.stringify(str, null, '  ');
@@ -63,7 +81,11 @@ function saveToFile(str, path){
 	
 	var f = new File(path);
 	
-	console.log('writing: file exists (overwriting)? '+f.exists());
+	if(isDebugOutput) console.log('writing: file exists (overwriting)? '+f.exists());
+	
+	if(doNotOverWrite && f.exists()){
+		return;//////////////////////// EARLY EXIT ///////////////////////
+	}
 	
 	var r = new FileWriter(f, false);
 	
@@ -71,5 +93,5 @@ function saveToFile(str, path){
 	
 	r.close();
 	
-	console.log('wrote String (len '+str.length+') to file: '+path);
+	if(isDebugOutput) console.log('wrote String (len '+str.length+') to file: '+path);
 }
